modified:   ../README.md 	modified:   ../TIDDIT.py 	modified:   DBSCAN.py 	modified:   TIDDIT.cpp 	new file:   TIDDIT_calling.py 	deleted:    TIDDIT_clustering.py 	new file:   TIDDIT_coverage.py 	new file:   TIDDIT_filtering.py 	new file:   TIDDIT_signals.py 	modified:   common.h 	modified:   data_structures/CoverageModule.cpp 	modified:   data_structures/ProgramModules.h 	modified:   data_structures/Translocation.cpp 	modified:   data_structures/Translocation.h 	modified:   data_structures/findTranslocationsOnTheFly.cpp 	modified:   setup.py
