+++ src/data_structures/CoverageModule.cpp
@@ -24,7 +24,7 @@ ostream& test(ofstream &coverageOutput,string output){
 
 }
 //constructor
-Cov::Cov(int binSize,string bamFile,string output,bool discordants){
+Cov::Cov(int binSize,string bamFile,string output){
 	ostream& covout=test(coverageOutput,output);
 	if(output != "stdout"){
 		coverageOutput.open((output+".tab").c_str());
@@ -42,18 +42,13 @@ Cov::Cov(int binSize,string bamFile,string output,bool discordants){
 	this -> binSize = binSize;
 	this -> currentChr=-1;
 	
-    this -> contigsNumber = 0;
-    this -> bamFile = bamFile;
-	this -> discordants=discordants;
+	this -> contigsNumber = 0;
+	this -> bamFile = bamFile;
 	this -> output = output;
 
-	if (this -> discordants == true){
-		covout << "#CHR" << "\t" << "start" << "\t" << "end" << "\t" << "coverage" <<"\t" << "quality" << "\t" << "discordants" << endl;
-	}else{
-		covout << "#CHR" << "\t" << "start" << "\t" << "end" << "\t" << "coverage" <<"\t" << "quality" << endl;
-	}
+	covout << "#CHR" << "\t" << "start" << "\t" << "end" << "\t" << "coverage" <<"\t" << "quality" << endl;
+	BamReader alignmentFile;
 
-    BamReader alignmentFile;
 	//open the bam file
 	alignmentFile.Open(bamFile);
 	
@@ -66,20 +61,14 @@ Cov::Cov(int binSize,string bamFile,string output,bool discordants){
 		contigsNumber++;
 	}
 	this -> coverageStructure.resize(contigsNumber);
-	this -> discordantsStructure.resize(contigsNumber);
 	this -> qualityStructure.resize(2);
 
 	qualityStructure[0].resize(contigsNumber);
 	qualityStructure[1].resize(contigsNumber);
 	for(int i=0;i<contigsNumber;i++){
-		//cout << i << endl;
-		//cout << ceil(contigLength[i]/double(binSize)) << endl;
 		coverageStructure[i].resize(ceil(contigLength[i]/double(binSize)),0);
 		qualityStructure[0][i].resize(ceil(contigLength[i]/double(binSize)),0);
 		qualityStructure[1][i].resize(ceil(contigLength[i]/double(binSize)),0);
-		if (this -> discordants == true){
-			discordantsStructure[i].resize(ceil(contigLength[i]/double(binSize)),0);
-		}
 	}
 	
 	
@@ -90,19 +79,15 @@ Cov::Cov(int binSize,string bamFile,string output,bool discordants){
 
 
 //this function calculates the coverage in bins of size binSize across the entire bamfile
-void Cov::bin(BamAlignment currentRead){
+void Cov::bin(BamAlignment currentRead, readStatus alignmentStatus){
 	//initialise the chromosome ID
 	if(currentChr == -1){
 		currentChr=currentRead.RefID;
 	}
 	
 	//if the quality of the read is high enough, it will be added to the data structure
-	readStatus alignmentStatus = computeReadType(currentRead, 100000,100, true);
 	if(alignmentStatus != lowQualty and alignmentStatus != unmapped) {
 		int element=floor(double(currentRead.Position)/double(binSize));
-		if (this -> discordants == true and currentRead.RefID != currentRead.MateRefID){
-			discordantsStructure[currentRead.RefID][element] +=1;
-		}
 		//if the entire read is inside the region, add all the bases to sequenced bases
 		if(currentRead.Position >= element*binSize and currentRead.Position+currentRead.Length-1 <= (element+1)*binSize){
 			coverageStructure[currentRead.RefID][element]+=currentRead.Length;
@@ -142,25 +127,21 @@ void Cov::printCoverage(){
 		static ostream& covout=cout;
 	}
 
-    for(int i=0;i<contigsNumber;i++){
-        for(int j=0;j<coverageStructure[i].size();j++){
-            int binStart = j*binSize;
-            int binEnd=binStart+binSize;
+	for(int i=0;i<contigsNumber;i++){
+		for(int j=0;j<coverageStructure[i].size();j++){
+			int binStart = j*binSize;
+			int binEnd=binStart+binSize;
             
-            if(binEnd > contigLength[i]){
-						binEnd=contigLength[i];
+			if(binEnd > contigLength[i]){
+				binEnd=contigLength[i];
 			}
-            double coverage=double(coverageStructure[i][j])/double(binEnd-binStart);
-            double quality = 0;
-            if(double(qualityStructure[1][i][j]) > 0){
-            	quality=double(qualityStructure[0][i][j])/double(qualityStructure[1][i][j]);
-            }
-			if (this -> discordants == true){
-				covout << position2contig[i] << "\t" << binStart << "\t" << binEnd << "\t" << coverage << "\t" << quality << "\t" << discordantsStructure[i][j] << endl;
-			}else{
-            	covout << position2contig[i] << "\t" << binStart << "\t" << binEnd << "\t" << coverage << "\t" << quality << endl;
+			double coverage=double(coverageStructure[i][j])/double(binEnd-binStart);
+			double quality = 0;
+			if(double(qualityStructure[1][i][j]) > 0){
+				quality=double(qualityStructure[0][i][j])/double(qualityStructure[1][i][j]);
 			}
-        }
+			covout << position2contig[i] << "\t" << binStart << "\t" << binEnd << "\t" << coverage << "\t" << quality << endl;
+		}
 	}
 	
 }
