+++ src/setup.py
@@ -3,5 +3,5 @@ from Cython.Build import cythonize
 
 setup(
     name = "TIDDIT",
-    ext_modules = cythonize(['TIDDIT_clustering.py',"DBSCAN.py"]),
+    ext_modules = cythonize(['TIDDIT_calling.py',"TIDDIT_coverage.py","TIDDIT_filtering.py","TIDDIT_signals.py","DBSCAN.py"]),
 )
